Refactor outbox + remove domain event type
