+++ eventastic_postgres/tests/common/test_aggregate.rs
@@ -1,6 +1,6 @@
 use eventastic::aggregate::Aggregate;
 use eventastic::aggregate::SideEffect;
-use eventastic::event::Event;
+use eventastic::event::DomainEvent;
 use serde::Deserialize;
 use serde::Serialize;
 use thiserror::Error;
@@ -32,7 +32,8 @@ pub enum AccountEvent {
     },
 }
 
-impl Event<Uuid> for AccountEvent {
+impl DomainEvent for AccountEvent {
+    type EventId = Uuid;
     fn id(&self) -> &Uuid {
         match self {
             AccountEvent::Open { event_id, .. }
@@ -67,9 +68,9 @@ pub enum SideEffects {
 
 impl SideEffect for SideEffects {
     /// The type used to uniquely identify this side effect.
-    type Id = Uuid;
+    type SideEffectId = Uuid;
 
-    fn id(&self) -> &Self::Id {
+    fn id(&self) -> &Self::SideEffectId {
         match self {
             SideEffects::PublishMessage { id, .. } | SideEffects::SendEmail { id, .. } => id,
         }
@@ -88,9 +89,6 @@ impl Aggregate for Account {
     /// The type of Domain Events that interest this Aggregate.
     /// Usually, this type should be an `enum`.
     type DomainEvent = AccountEvent;
-
-    /// The type used to uniquely identify the a given domain event.
-    type DomainEventId = Uuid;
 
     /// The error type that can be returned by [`Aggregate::apply`] when
     /// mutating the Aggregate state.
